refactor(abstract-token-network-job): Removes user ID from AbsractTokenNetworkJob constructor as it is not used

+++ test/libsyncengine/jobs/network/kDrive_API/testapitranslator.cpp
@@ -65,7 +65,7 @@ void TestApiTranslator::setUp() {
 
     _driveDbId = 1;
     _driveId = atoi(testVariables.driveId.c_str());
-    Drive drive(static_cast<int>(_driveDbId), static_cast<int>(_driveId), account.dbId(), std::string(), 0, std::string());
+    Drive drive(_driveDbId, _driveId, account.dbId(), std::string(), 0, std::string());
     (void) ParmsDb::instance()->insertDrive(drive);
 
     // Setup proxy
